[WORKING] :First version of the planner working with KinoDyn.
Acado is used as a steering method with min time problem.
The PRM is comupted with a brute force "1000 neighbors" approaches.
A* compute the best sequence in the PRM.
Acado is used to optimize the path output by A*.
We plot the trajectory on the brackground of Value function computed by the neural net.
--- the sampling using ACnet + acado optim is not working, or seems not to. Why?

Lets try to do the same with a double pendulum.